fix: revert Transform rotation to Float

+++ packages/aztecs-transform/src/Data/Aztecs/Transform.hs
@@ -5,7 +5,7 @@ import Linear (V2 (..))
 
 data Transform = Transform
   { transformPosition :: V2 Int,
-    transformRotation :: Int,
+    transformRotation :: Float,
     transformScale :: V2 Int
   }
   deriving (Eq, Show)
